changed discord to widget on signupnow component
